# [SPARK-42528][CORE] Optimize PercentileHeap