--- a/core/src/main/scala/org/apache/spark/util/collection/PercentileHeap.scala
+++ b/core/src/main/scala/org/apache/spark/util/collection/PercentileHeap.scala
@@ -20,97 +20,55 @@ package org.apache.spark.util.collection
 import scala.collection.mutable.PriorityQueue
 
 /**
- * PercentileHeap is designed to be used to quickly track the percentile of a group of numbers
- * that may contain duplicates. Inserting a new number has O(log n) time complexity and
- * determining the percentile has O(1) time complexity.
- * The basic idea is to maintain two heaps: a smallerHalf and a largerHalf. The smallerHalf
- * stores the smaller half of all numbers while the largerHalf stores the larger half.
- * The sizes of two heaps need to match the percentage each time when a new number is inserted so
- * that the ratio of their sizes is percentage to (1 - percentage). Therefore each time when
- * percentile() is called we check if the sizes of two heaps match the percentage. If they do,
- * we should return the average of the two top values of heaps. Otherwise we return the top of the
- * heap which exceeds its percentage.
+ * PercentileHeap tracks the percentile of a collection of numbers.
+ *
+ * Insertion is O(log n), Lookup is O(1).
+ *
+ * The implementation keeps two heaps: a bottom heap (`botHeap`) and a top heap (`topHeap`). The
+ * bottom heap stores all the numbers below the percentile and the top heap the ones above the
+ * percentile. During insertion the relative sizes of the heaps are adjusted to match the
+ * target percentile.
  */
-private[spark] class PercentileHeap(percentage: Double = 0.5)(implicit val ord: Ordering[Double]) {
-  assert(percentage >= 0 && percentage <= 1)
+private[spark] class PercentileHeap(percentage: Double = 0.5) {
+  assert(percentage > 0 && percentage < 1)
+
+  private[this] val topHeap = PriorityQueue.empty[Double](Ordering[Double].reverse)
+  private[this] val botHeap = PriorityQueue.empty[Double](Ordering[Double])
+
+  def isEmpty(): Boolean = botHeap.isEmpty && topHeap.isEmpty
+
+  def size(): Int = botHeap.size + topHeap.size
 
   /**
-   * Stores all the numbers less than the current percentile in a smallerHalf,
-   * i.e percentile is the maximum, at the root.
+   * Returns percentile of the inserted elements as if the inserted elements were sorted and we
+   * returned `sorted((sorted.length * percentage).toInt)`.
    */
-  private[this] val smallerHalf = PriorityQueue.empty[Double](ord)
-
-  /**
-   * Stores all the numbers greater than the current percentile in a largerHalf,
-   * i.e percentile is the minimum, at the root.
-   */
-  private[this] val largerHalf = PriorityQueue.empty[Double](ord.reverse)
-
-  def isEmpty(): Boolean = {
-    smallerHalf.isEmpty && largerHalf.isEmpty
+  def percentile(): Double = {
+    if (isEmpty) throw new IndexOutOfBoundsException("empty")
+    topHeap.head
   }
-
-  def size(): Int = {
-    smallerHalf.size + largerHalf.size
-  }
-
-  // Exposed for testing.
-  def smallerSize(): Int = smallerHalf.size
 
   def insert(x: Double): Unit = {
-    // If both heaps are empty, we insert it to the heap that has larger percentage.
     if (isEmpty) {
-      if (percentage < 0.5) smallerHalf.enqueue(x) else largerHalf.enqueue(x)
+      topHeap.enqueue(x)
     } else {
-      // If the number is larger than current percentile, it should be inserted into largerHalf,
-      // otherwise smallerHalf.
-      if (x > percentile) {
-        largerHalf.enqueue(x)
+      val p = topHeap.head
+      val growBot = ((size + 1) * percentage).toInt > botHeap.size
+      if (growBot) {
+        if (x < p) {
+          botHeap.enqueue(x)
+        } else {
+          topHeap.enqueue(x)
+          botHeap.enqueue(topHeap.dequeue)
+        }
       } else {
-        smallerHalf.enqueue(x)
+        if (x < p) {
+          botHeap.enqueue(x)
+          topHeap.enqueue(botHeap.dequeue())
+        } else {
+          topHeap.enqueue(x)
+        }
       }
-    }
-    rebalance()
-  }
-
-  // Calculate the deviation between the ratio of smaller heap size to larger heap size and the
-  // expected ratio, which is percentage : (1 - percentage). Negative result means the smaller
-  // heap has too less elements, positive result means the smaller heap has too many elements.
-  private def calculateDeviation(smallerSize: Int, largerSize: Int): Double = {
-    smallerSize * (1 - percentage) - largerSize * percentage
-  }
-
-  private[this] def rebalance(): Unit = {
-    // If moving one value from heap to the other heap can fit the percentage better, then
-    // move it.
-    val currentDev = calculateDeviation(smallerHalf.size, largerHalf.size)
-    if (currentDev > 0) {
-      val newDev = calculateDeviation(smallerHalf.size - 1, largerHalf.size + 1)
-      if (math.abs(newDev) < currentDev) {
-        largerHalf.enqueue(smallerHalf.dequeue)
-      }
-    }
-    if (currentDev < 0) {
-      val newDev = calculateDeviation(smallerHalf.size + 1, largerHalf.size - 1)
-      if (math.abs(newDev) < -currentDev) {
-        smallerHalf.enqueue(largerHalf.dequeue())
-      }
-    }
-  }
-
-  def percentile: Double = {
-    if (isEmpty) {
-      throw new NoSuchElementException("PercentileHeap is empty.")
-    }
-    val dev = calculateDeviation(smallerHalf.size, largerHalf.size)
-    // If the deviation is very small, we take the average of the top elements from the two heaps
-    // as the percentile.
-    if (smallerHalf.nonEmpty && largerHalf.nonEmpty && math.abs(dev / size) < 0.01) {
-      (largerHalf.head + smallerHalf.head) / 2.0
-    } else if (dev < 0) {
-      largerHalf.head
-    } else {
-      smallerHalf.head
     }
   }
 }