@alkis 2023-02-23
> Also given this is an optimization change - include benchmark to quantify the impact ?

I did benchmarking live in a cluster. Profiles before show ~1% of scheduler time in PercentileHeap operations. Profiles after do not have PercentileHeap operations at all.