@alkis 2023-02-23
> Can we minimize diff's to this file ? A large fraction is whitespace changes and due to the renames ... will take a look at the changes as well.

Can you treat is a new implementation? There is only 15 lines (L55-L70) that matter on the new implementation - the code inside `insert`. Outside of `insert` there is nothing meaty or interesting.