@alkis 2023-02-23
I updated the implementation and the description. TLDR I use a comparator-less java PriorityQueue now for a total of 100x speedup over the original implementation.